# openapi.yaml
openapi: 3.0.0
info:
  title: Todo API
  description: A RESTful API for managing todo items
  version: 1.0.0
servers:
  - url: http://127.0.0.1:5000
    description: Local development server
paths:
  /todos:
    get:
      summary: Get all todos
      security:
        - apiKey: []
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Todo'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
    post:
      summary: Create a new todo
      security:
        - apiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: Title of the todo
                  required: true
                description:
                  type: string
                  description: Description of the todo
                completed:
                  type: boolean
                  description: Whether the todo is completed
                  default: false
      responses:
        '201':
          description: Todo created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /todos/{todo_id}:
    get:
      summary: Get a specific todo
      security:
        - apiKey: []
      parameters:
        - name: todo_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '404':
          description: Todo not found
    put:
      summary: Update a todo
      security:
        - apiKey: []
      parameters:
        - name: todo_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: Title of the todo
                description:
                  type: string
                  description: Description of the todo
                completed:
                  type: boolean
                  description: Whether the todo is completed
      responses:
        '200':
          description: Todo updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '404':
          description: Todo not found
    delete:
      summary: Delete a todo
      security:
        - apiKey: []
      parameters:
        - name: todo_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Todo deleted successfully
        '404':
          description: Todo not found
components:
  schemas:
    Todo:
      type: object
      properties:
        id:
          type: integer
          description: The unique identifier of the todo
        title:
          type: string
          description: Title of the todo
        description:
          type: string
          description: Description of the todo
        completed:
          type: boolean
          description: Whether the todo is completed
        created_at:
          type: string
          format: date-time
          description: When the todo was created
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key